example config creator clap option Fixes #32
