+++ docs/src/configuration.md
@@ -2,6 +2,9 @@
 
 I'll totally forget to update this, so check [the rustdoc for ConfigFile](https://goatns.dotgoat.net/rustdoc/goatns/config/struct.ConfigFile.html) for up-to-date information.
 
+To write out a "default" config file, run `goatns --export-default-config` which will dump the contents of the system defaults.
+
+
 ## User Authentication
 
 This is build for [Kanidm](https://kanidm.com) but should work with any OIDC identity provider.
