+++ src/utils.rs
@@ -1,17 +1,6 @@
-use crate::datastore::Command;
-use crate::enums::SystemState;
-use crate::zones::FileZone;
 use crate::HEADER_BYTES;
-use clap::{arg, command, value_parser, Arg, ArgMatches};
-use dialoguer::theme::ColorfulTheme;
-use dialoguer::Confirm;
-use dialoguer::Input;
 use log::{debug, trace};
 use std::str::from_utf8;
-use tokio::io::AsyncWriteExt;
-use tokio::sync::mpsc;
-use tokio::sync::oneshot;
-use tokio::time::sleep;
 
 pub fn vec_find(item: u8, search: &[u8]) -> Option<usize> {
     for (index, curr_byte) in search.iter().enumerate() {
@@ -264,65 +253,6 @@ pub fn hexdump(bytes: Vec<u8>) {
     }
 }
 
-pub fn clap_parser() -> ArgMatches {
-    command!() // requires `cargo` feature
-        // .arg(arg!([name] "Optional name to operate on"))
-        .arg(
-            arg!(
-                -c --config <FILE> "Sets a custom config file"
-            )
-            .required(false)
-            .value_parser(value_parser!(String)),
-        )
-        .arg(
-            Arg::new("configcheck")
-                .short('t')
-                .long("configcheck")
-                .help("Check the config file, show it and then quit.")
-                .action(clap::ArgAction::SetTrue),
-        )
-        .arg(
-            Arg::new("export_zone")
-                .short('e')
-                .long("export-zone")
-                .help("Export a single zone.")
-                .value_parser(value_parser!(String)),
-        )
-        .arg(
-            Arg::new("import_zones")
-                .short('i')
-                .long("import-zones")
-                .help("Import a single zone file.")
-                .action(clap::ArgAction::SetTrue),
-        )
-        .arg(
-            Arg::new("import_zone")
-                .long("import-zone")
-                .help("Import a single zone from a file.")
-                .value_parser(value_parser!(String)),
-        )
-        .arg(
-            Arg::new("filename")
-                .short('f')
-                .long("filename")
-                .help("Filename to save to (used in other commands).")
-                .value_parser(value_parser!(String)),
-        )
-        .arg(
-            Arg::new("add_admin")
-                .long("add-admin")
-                .help("Add a new admin user.")
-                .action(clap::ArgAction::SetTrue),
-        )
-        .arg(
-            Arg::new("use_zonefile")
-                .long("using-zonefile")
-                .help("Load the zone file into the DB on startup, typically used for testing.")
-                .action(clap::ArgAction::SetTrue),
-        )
-        .get_matches()
-}
-
 /// turn a degrees/minutes/seconds format into unsigned 32-bit integer matching the format
 /// required for a DNS LOC record
 ///
@@ -364,205 +294,4 @@ pub fn loc_size_to_u8(input: f32) -> u8 {
     // turn it into the magic ugly numbers
     let retval: u8 = (mantissa << 4) | (exponent as u8);
     retval
-}
-
-pub async fn add_admin_user(tx: mpsc::Sender<Command>) -> Result<(), ()> {
-    // prompt for the username
-    println!("Creating admin user, please enter their username from the identity provider");
-    let username: String = Input::with_theme(&ColorfulTheme::default())
-        .with_prompt("Username")
-        .interact_text()
-        .unwrap();
-
-    println!(
-        "The authentication reference is the unique user identifier in the Identity Provider."
-    );
-    let authref: String = Input::with_theme(&ColorfulTheme::default())
-        .with_prompt("Authentication Reference:")
-        .interact_text()
-        .unwrap();
-
-    println!(
-        r#"
-
-Creating the following user:
-
-
-Username: {username}
-Authref:  {authref}
-
-"#
-    );
-    // show the details and confirm them
-    let confirm = Confirm::with_theme(&ColorfulTheme::default())
-        .with_prompt("Do these details look correct?")
-        .interact_opt();
-    if confirm.is_err() || confirm.unwrap().is_none() {
-        log::warn!("Cancelled user creation");
-        return Err(());
-    };
-
-    // create oneshot
-    let (tx_oneshot, rx_oneshot) = oneshot::channel();
-
-    let new_user = Command::CreateUser {
-        username: username.clone(),
-        authref: authref.clone(),
-        admin: true,
-        disabled: false,
-        resp: tx_oneshot,
-    };
-    // send command
-    if let Err(error) = tx.send(new_user).await {
-        log::error!("Failed to send new user command for username {username:?}: {error:?}");
-        return Err(());
-    };
-    // wait for the response
-    match rx_oneshot.await {
-        Ok(res) => match res {
-            true => {
-                log::info!("Successfully created user!");
-                Ok(())
-            }
-            false => {
-                log::error!("Failed to create user! Check datastore logs.");
-                Err(())
-            }
-        },
-        Err(error) => {
-            log::debug!("Failed to rx result from datastore: {error:?}");
-            Err(())
-        }
-    }
-}
-
-pub async fn cli_commands(
-    tx: mpsc::Sender<Command>,
-    clap_results: &ArgMatches,
-) -> Result<SystemState, String> {
-    if clap_results.get_flag("add_admin") {
-        let _ = add_admin_user(tx).await;
-
-        return Ok(SystemState::ShuttingDown);
-    }
-
-    if let Some(zone_name) = clap_results.get_one::<String>("export_zone") {
-        if let Some(output_filename) = clap_results.get_one::<String>("filename") {
-            log::info!("Exporting zone {zone_name} to {output_filename}");
-            let res = export_zone_file(tx, zone_name, output_filename).await;
-            if let Err(err) = res {
-                log::error!("{err}");
-            }
-            return Ok(SystemState::Export);
-        } else {
-            log::error!("You need to specify a a filename to save to.");
-            return Ok(SystemState::ShuttingDown);
-        }
-    };
-
-    if clap_results.get_flag("import_zones") {
-        if let Some(filename) = clap_results.get_one::<String>("filename") {
-            log::info!("Importing zones from {filename}");
-            import_zones(tx, filename.to_owned(), None)
-                .await
-                .map_err(|e| format!("Error importing {filename}: {e:?}"))?;
-
-            return Ok(SystemState::Import);
-        } else {
-            log::error!("You need to specify a a filename to save to.");
-            return Ok(SystemState::ShuttingDown);
-        }
-    };
-    if let Some(zone_name) = clap_results.get_one::<String>("import_zone") {
-        if let Some(filename) = clap_results.get_one::<String>("filename") {
-            log::info!("Importing zones from {filename}");
-            import_zones(tx, filename.to_owned(), Some(zone_name.to_owned()))
-                .await
-                .map_err(|e| format!("Error importing {filename}: {e:?}"))?;
-
-            return Ok(SystemState::Import);
-        } else {
-            log::error!("You need to specify a a filename to save to.");
-            return Ok(SystemState::ShuttingDown);
-        }
-    };
-    Ok(SystemState::Server)
-}
-/// dump a zone to a file
-pub async fn export_zone_file(
-    tx: mpsc::Sender<Command>,
-    zone_name: &String,
-    filename: &String,
-) -> Result<(), String> {
-    // make a channel
-
-    let (tx_oneshot, rx_oneshot) = oneshot::channel();
-    let ds_req: Command = Command::GetZone {
-        id: None,
-        name: Some(zone_name.clone()),
-        resp: tx_oneshot,
-    };
-    if let Err(error) = tx.send(ds_req).await {
-        return Err(format!(
-            "failed to send to datastore from export_zone_file {error:?}"
-        ));
-    };
-
-    let zone: Option<FileZone> = match rx_oneshot.await {
-        Ok(value) => value,
-        Err(err) => return Err(format!("rx from ds failed {err:?}")),
-    };
-    eprintln!("Got filezone: {zone:?}");
-
-    if zone.is_none() {
-        log::warn!("Couldn't find the zone {zone_name}");
-        return Ok(());
-    }
-
-    // dump the zone
-    let zone_bytes = serde_json::to_string_pretty(&zone.unwrap()).unwrap();
-
-    // open the file
-    let mut file = tokio::fs::File::create(filename)
-        .await
-        .map_err(|e| format!("Failed to open file {e:?}"))
-        .unwrap();
-    // write the thing
-    file.write_all(zone_bytes.as_bytes())
-        .await
-        .map_err(|e| format!("Failed to write file: {e:?}"))
-        .unwrap();
-    // make some cake
-
-    Ok(())
-}
-
-pub async fn import_zones(
-    tx: mpsc::Sender<Command>,
-    filename: String,
-    zone_name: Option<String>,
-) -> Result<(), String> {
-    let (tx_oneshot, mut rx_oneshot) = oneshot::channel();
-    let msg = Command::ImportFile {
-        filename,
-        resp: tx_oneshot,
-        zone_name,
-    };
-    if let Err(err) = tx.send(msg).await {
-        log::error!("Failed to send message to datastore: {err:?}");
-    }
-    loop {
-        let res = rx_oneshot.try_recv();
-        match res {
-            Err(error) => {
-                if let oneshot::error::TryRecvError::Closed = error {
-                    break;
-                }
-            }
-            Ok(()) => break,
-        };
-        sleep(std::time::Duration::from_micros(500)).await;
-    }
-    Ok(())
-    // rx_oneshot.await.map_err(|e| format!("Failed to receive result: {e:?}"))
 }
